requirements.txt
Flask
Pillow
requests
pytz
beautifulsoup4
tyme4py
qrcode